==> R/temperature_degree_c/1.2_explore_temp_filtered_data.qmd <==
---
format: 
  html:
    toc: true
    toc-depth: 3
    embed-resources: true
editor_options: 
  chunk_output_type: console
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, dpi = 600, fig.width = 8)

library(dplyr)
library(DT)
library(here)
library(ggplot2)
library(leaflet)
library(lubridate)
library(plotly)
library(RColorBrewer)
library(readr)
library(summaryplots)

source(here("functions/subchunkify.R"))

theme_set(theme_light())

dt_options <- list(
  dom = 'ft',
  paging = FALSE,
  searching = TRUE,
  scrollY = "500px",
  columnDefs = list(list(className = 'dt-center', targets = "_all"))
)

# summarized data 
dat <- read_csv(
  here("data/2_summary_filtered_data.csv"), show_col_types = FALSE
) %>% 
  filter(variable == "Temperature")

# station locations
st_locations <- read_csv(
  here("data/Station_Locations_2022-12-06.csv"), show_col_types = FALSE
)

county_pal <- get_county_colour_palette(length(unique(st_locations$COUNTY)))

theme_facet_plotly <- theme(panel.spacing.y = unit(15, "lines"))
```

# EXPLORATORY ANALYSIS: TEMPERATURE

July 26, 2023

- Data submitted to the Open Data Portal in December 2022.
- Depth rounded to nearest whole number
  - Results in depth of 0 m for 1042 (Annapolis), Captains Pond (Antigonish), Ile due Havre (Antigonish), Sober Island (Halifax), and Back Harbour (Pictou).
- Preliminary QC was applied to the data.
  - Obvious outliers removed.
- Additional QC applied was based on review of 1.1_explore_temp_data.html
  - Fresh water and other "outlier" stations not considered.
    - Piper Lake, Hourglass Lake, 0193 (Denas Pond), Sissiboo
    - Data below the thermocline in Whycocomagh Bay 
      - 0814x East and 0814x West, Whycocomagh Basin: 18 m
      - Aberdeen, Whycocomagh Bay: 18, 23, and 28 m
      - Deep Basin, Whycocomagh Basin: 26, 36, and 40 m

    
# Station Locations

Approximate location of stations with temperature data.
Marker size is proportional to the number of temperature observations within the county.

```{r}
#| fig-height: 7

# join the station locations dataset with the number of obs from each station
st_locations <- st_locations %>% 
  inner_join(
    dat %>% 
      filter(group == "all_station") %>% 
      select(COUNTY = county, STATION = station, n),
    by = c("COUNTY", "STATION") 
  ) %>% 
  group_by(COUNTY) %>% 
  mutate(
    n_prop = round(n / sum(n), digits = 2),
    popup = paste(COUNTY, STATION, n_prop, sep = "</br>")
  )  

# interactive map
leaflet(st_locations) %>% 
  addProviderTiles(providers$CartoDB.Positron) %>% 
  add_circle_markers_county(
    county_pal = county_pal, 
    size =  ~n_prop * 25
  ) %>% 
  add_county_fill_legend(county_pal = county_pal)
```

# Gross Range

- These means are likely influenced by the timing of the observations (see figures 14 - 18).

```{r}
county <- dat %>% 
  filter(group == "county" | group == "all_data") %>% 
  mutate(
    county = if_else(group == "all_data", "All", county),
    n_percent = round(n * 100 / max(n), digits = 1)
  ) %>% 
  select(-c(year, depth, month, station))
```

## County

::: panel-tabset

### Figure 1
```{r}
#| fig-height: 8
#| fig-cap: The number of temperatures observations in each county. The number at the end of each bar indicates the percent of the total temperature observations from the associated county.
#| fig-cap-location: top

p <- county  %>% 
  filter(county != "All") %>% 
  plot_n_obs_county()

ggplotly(p)

```

### Figure 2
```{r}
#| fig-height: 8
#| fig-cap: The mean and standard deviation of temperature in each county.
#| fig-cap-location: top

p <- plot_mean_sd_county(county)

ggplotly(p, tooltip = "text")
```

### Table 1
```{r}
#| fig-cap: Summary statistics of temperature observations - total and by county. Grossrange user-defined thresholds are calculated as user_min = mean - 3 * stdev and user_max = mean + 3 * stdev.
#| fig-cap-location: top

county %>% 
  select(county, n, mean, stdev) %>% 
  mutate(user_min = mean - 3 * stdev, user_max = mean + 3 * stdev) %>% 
  datatable(options = dt_options, rownames = FALSE)
```
:::

## Depth

```{r}
depth <- dat %>% 
  filter(group == "all_depth") %>% 
  mutate(n_percent = round(n * 100 / sum(n), digits = 1)) %>% 
  select(-c(month, group, station, units, county))
```

::: panel-tabset

### Figure 3
```{r}
#| fig-height: 10
#| fig-cap: Number of temperature observations at each depth.
#| fig-cap-location: top

p <- plot_n_obs_depth(depth, facet_county = FALSE)

ggplotly(p, tooltip = "text")
```

### Figure 4
```{r}
#| fig-height: 10
#| fig-cap: Mean and standard deviation of temperature at each depth.
#| fig-cap-location: top

p <- plot_mean_sd_depth(depth, facet_county = FALSE, text_size = 12)

ggplotly(p, tooltip = "text")

```
:::

## County + Depth

```{r}
county_depth <- dat %>% filter(group == "county_depth") 
```

::: panel-tabset

### Figure 5
```{r}
#| fig.height: 8
#| fig-cap: Depths with temperature measurements by county.
#| fig-cap-location: top

p <- ggplot(county_depth, aes(depth, county, col = factor(depth))) +
  geom_point(size = 4) +
  scale_y_discrete(name = "", limits = rev) +
  scale_x_continuous("Sensor Depth Below the Surface at Low Tide (m)")

ggplotly(p)
```

### Figure 6
```{r}
#| fig-height: 10
#| fig-cap: The number of temperature obervations at each depth for each county.
#| fig-cap-location: top

p <- plot_n_obs_depth(county_depth) +
   theme_facet_plotly

ggplotly(p)
```

### Figure 7
```{r}
#| fig-height: 10
#| fig-cap: Average temperature and standard deviation for each county at each measured depth. 

p <- plot_mean_sd_depth(county_depth, text_size = 12) +
   theme_facet_plotly

ggplotly(p, tooltip = "text")
```

### Figure 8

```{r}
#| fig-height: 8
#| fig-cap: Average temperature and standard deviation of all observations (grey) and average temperature at each depth.

p <- ggplot(filter(county, county != "All"), aes(mean, county)) +
  geom_errorbar(
    aes(xmin = mean - stdev, xmax = mean + stdev), 
    width = 0, linewidth = 1, col = "grey70"
  ) +
  geom_point(size = 5, pch = 21, fill = "grey70", col = "grey50") +
  geom_point(
    data = county_depth, 
    aes(mean, county, col = factor(depth)), 
    alpha = 0.75, size = 2
  ) +
  scale_y_discrete(name = "", limits = rev) +
  scale_x_continuous("Temperature (degree C)")

ggplotly(p)
```
:::

- Digby 
  - 31 m: measured at Long Island from January - May (cold months).
  - 32 m: measured at Long Island 2 from July - November (warm months).
  - Monthly climatology for Digby looks reasonable (Figure 15, 16).
  - Removing these data series has a negligible impact on the distribution of observations, mean, standard deviation, and user thresholds (not shown).
- Guysborough 
  - 11 m: measured at Tanner Island and English Harbour in February - March 2019 (cold months). This is the near-bottom temperature for both stations.
  - Monthly climatology for Guysborough looks reasonable (Figure 15, 16).
- Inverness: only one deployment for many depths.
- Lunenburg: clear decrease in temperature at depth.
- Shelburne
  - 16 m: measured at Ingomar from January - May (cold months), then sensor moved to 15 m
- Victoria: not a lot of difference in temperature through the water column.

# Climatology

## Month
```{r}
month <- dat %>% 
  filter(group == "all_month") %>% 
  mutate(
    n_percent = round(n * 100 / sum(n), digits = 1),
    month = month(month, label = TRUE)
  ) 
```

::: panel-tabset

### Figure 9
```{r}
#| fig-height: 8
#| fig-cap: The number of temperatures observations in each month. The number at the end of each bar indicates the percent of the total temperature observations from the associated month.
#| fig-cap-location: top

p <- plot_n_obs_month(month, facet_county = FALSE)

ggplotly(p)
```

### Figure 10
```{r}
#| fig-height: 8
#| fig-cap-location: top

p <- plot_mean_sd_season(month, facet_county = FALSE)

ggplotly(p)

```

### Table 2
```{r}
#| fig-cap: Summary statistics of temperature observations.
#| fig-cap-location: top

month %>% 
  select(month, n, mean, stdev) %>% 
  mutate(season_min = mean - 3 * stdev, season_max = mean + 3 * stdev) %>% 
  datatable(options = dt_options, rownames = FALSE)
```
:::

## Month + Year 

```{r}
month_year <- dat %>% 
  filter(group == "all_month_year") %>%
  group_by(year) %>% 
  mutate(
    n_percent = round(n * 100 / sum(n), digits = 1),
    month = month(month, label = TRUE)
  ) 
```

::: panel-tabset

### Figure 11
```{r}
#| fig-height: 10
#| fig-cap: The number of temperature observations for each month by year.
#| fig-cap-location: top

p <- plot_n_obs_month(month_year, text_size = 12) +
  facet_wrap(~year) +
  theme(panel.spacing.y = unit(15, "lines"))

ggplotly(p)
```

### Figure 12
```{r}
#| fig-height: 10
#| fig-cap: Monthly climatology by year.
#| fig-cap-location: top

p <- plot_mean_sd_season(month_year) +
  facet_wrap(~year) +
  theme(panel.spacing.y = unit(15, "lines"))

ggplotly(p)
```

### Figure 13
```{r}
#| fig-height: 8
#| fig-cap: Monthly climatology by year - interactive.
#| fig-cap-location: top

month_year <- month_year %>% 
  mutate(year = factor(year))

p <- plot_mean_sd_season_ribbon(
  month_year, group = "year", facet_county = FALSE
)

ggplotly(p, tooltip = "text")

```

### Table 3
```{r}
month_year %>% 
  select(year, month, n, mean, stdev) %>% 
  datatable(options = dt_options, rownames = FALSE)
```

:::

- Consistent observations starting in 2018
- Fairly uniform coverage for each month (except for January 2018)
- 2019 was the coldest year (in February + March)
- 2017 was relatively warm

## Month + County
```{r}
county_month <- dat %>% 
  filter(group == "county_month") %>% 
  mutate(month = month(month, label = TRUE)) %>% 
  select(-c(units, group, year, station, depth))
```

::: panel-tabset

### Figure 14
```{r}
#| fig-height: 12
#| fig-cap: The number of temperature observations in each month for each county.
#| fig-cap-location: top

p <- plot_n_obs_month(county_month, text_size = 12) +
  theme(panel.spacing.y = unit(15, "lines"))

ggplotly(p)

```

### Figure 15
```{r}
#| fig-height: 10
#| fig-cap: The mean and standard deviation of temperature in each county.
#| fig-cap-location: top

p <- plot_mean_sd_season(county_month) +
  theme(panel.spacing.y = unit(15, "lines"))

ggplotly(p)
```

### Figure 16
```{r}
#| fig-height: 8
#| fig-cap: The mean and standard deviation of temperature in each county.
#| fig-cap-location: top

p <- plot_mean_sd_season_ribbon(
  county_month, facet_county = FALSE, group = "county", text_size = 12
)

ggplotly(p, tooltip = "text")

```

### Figure 17
```{r}
#| fig-height: 8
#| fig-cap: The mean and standard deviation of temperature in each county.
#| fig-cap-location: top

county_gr <- county_month %>% 
  rename(mean_month = mean) %>% 
  group_by(county) %>%
  summarise(
    mean = round(mean(mean_month), digits = 3),
    stdev = round(sd(mean_month), digits = 3),
    group = "county_climatology"
  ) %>% 
  bind_rows(
    county %>% 
      filter(group != "all_data") %>% 
      select(-c(n, n_percent))
  ) 

p <- ggplot(county_gr, aes(mean, county, fill = group, col = group)) +
  geom_errorbar(
    aes(xmin = mean - stdev, xmax = mean + stdev), 
    width = 0, linewidth = 1, alpha = 0.75
  ) +
  geom_point(size = 3, pch = 21, col = 1) +
  scale_y_discrete(name = "", limits = rev) +
  scale_x_continuous("Temperature (degree C)")

ggplotly(p)
```

### Figure 18
```{r}
#| fig-height: 8
#| fig-cap: The mean and standard deviation of temperature in each county.
#| fig-cap-location: top

p <- ggplot(county_gr, aes(stdev, county, fill = group)) +
  geom_col(position = "dodge") +
  scale_y_discrete(name = "", limits = rev) +
  scale_x_continuous("Standard Deviation (degree C)")

ggplotly(p)
```

### Figure 19
```{r}
#| label: fig-temp-stdev
#| fig-height: 9
#| fig-cap: The standard deviation of temperature for each calendar month in each county. The shaded yellow region indicates "summer" months (June, July, August, and September). The red bar indicates the highest standard deviation for each county.

county_month %>% 
  group_by(county) %>% 
  mutate(
    max_stdev = max(stdev),
    max_stdev = if_else(max_stdev == stdev, 1, 0)
  ) %>% 
  ggplot(aes(month, stdev, fill = factor(max_stdev))) +
  geom_col() + # need to put this here or the months are displayed alphabetically
  scale_fill_manual(values = c("#8DD3C7", "#FB8072")) +
  annotate(
    "rect", xmin = 5.5, xmax = 9.5, ymin = -Inf,  ymax = Inf,
    col = "#FFFF8A", fill = "#FFFF8A", alpha = 0.5
  ) +
  geom_col() +
  scale_x_discrete(name = "", breaks = c("Jan", "Mar", "May", "Jul", "Sep", "Nov")) +
  scale_y_continuous(
    "Standard Deviation (degrees C)",
    expand = expansion(mult = c(0, .1))
  ) +
  facet_wrap(~county, ncol = 3) +
  theme(legend.position = "none")

```
:::

- The *overall* mean is influenced by the number of observations each month
  - e.g., Inverness: most temperature observations from June to November (Figure 14). So the overall mean is substantially lower when calculated from the climatology (vs. from the raw observations; Figure 17).
    - similar for Antigonish, Pictou, and Yarmouth
    - Cape Breton and Colchester have more observations in the Winter, and so the overall mean is *higher* when calculated from the climatology.
    

## Month + County + Year
```{r}
county_month_year <- dat %>% 
  filter(group == "county_month_year") %>% 
  group_by(year, month) %>% 
  mutate(
    n_percent = round(n * 100 / sum(n), digits = 1),
    month = month(month, label = TRUE)
  ) %>% 
  select(-c(units, group, station, depth)) %>% 
  ungroup()
```

### Number of Obervations

Note: x-axis is not consistent between figures.

Figures are not interactive.

::: panel-tabset

```{r}
#| results: "asis"

counties <- unique(county_month_year$county)

for(i in seq_along(counties)){
  
  county_i <- counties[i]
  dat_i <- county_month_year %>% filter(county == county_i)
  
  # height of figure j
  n_years <- length(unique(dat_i$year))
  if(n_years <= 3) h = 4
  if(n_years > 3) h = 8
  
  cat('\n###', county_i, '\n')
  
  p <- plot_n_obs_month(dat_i, text_size = 12) +
    facet_wrap(~year, ncol = 3) 
  
  subchunkify(p, fig_height = h, fig_width = 8.5)
}
```
:::

### Climatology

Figures are not interactive.

::: panel-tabset

```{r}
#| fig-height: 6
#| results: "asis"

n_years <- length(unique(county_month_year$year))

county_month_year <- county_month_year %>% 
  mutate(year = factor(year))

for(i in seq_along(counties)){
  
  county.i <- counties[i]
  
  dat.i <- county_month_year %>% 
    filter(county == county.i) 
  
  cat('\n###', county.i, '\n')
  
  p <- plot_mean_sd_season_ribbon(
    dat.i, group = "year",
    n_group = n_years,
    facet_county = FALSE, y_limits = c(-2, 27)
  )

  print(p)
}
  
```
:::

## Month + County + Depth
```{r}
county_month_depth <- dat %>% 
  filter(group == "county_month_depth") %>% 
  group_by(year, month) %>% 
  mutate(
    n_percent = round(n * 100 / sum(n), digits = 1),
    month = month(month, label = TRUE)
  ) %>% 
  select(-c(variable, units, group, station)) %>% 
  ungroup()
```

### Number of Obervations

Note: x-axis is not consistent between figures.

Figures are not interactive.

::: panel-tabset

```{r}
#| results: "asis"

counties <- unique(county_month_depth$county)

for(i in seq_along(counties)){
  
  county_i <- counties[i]
  dat_i <- county_month_depth %>% filter(county == county_i)
  
  # height of figure j
  n_depth <- length(unique(dat_i$depth))
  if(n_years <= 3) h = 4
  if(n_years > 3) h = 8
  
  cat('\n###', county_i, '\n')
  
  p <- plot_n_obs_month(dat_i, text_size = 12) +
    facet_wrap(~depth, ncol = 3) 
  
  subchunkify(p, fig_height = h, fig_width = 8.5)
}
```
:::

```{r}
zero_obs <- county_month_depth %>% 
  distinct(month, county) %>% 
  full_join(distinct(county_month_depth, county, depth), by = "county") %>% 
  full_join(county_month_depth, by = c("month", "county", "depth")) %>% 
  select(-c(mean, stdev, year, n_percent)) %>% 
  mutate(n = if_else(is.na(n), 0, n)) %>% 
  filter(n == 0) 
```

There are `r nrow(zero_obs)` month-county-depth combinations without any observations.

### Table 4

Month-county-depth combinations with zero observations.

```{r}
zero_obs %>% 
  select(-n) %>% 
  datatable(options = dt_options, rownames = FALSE)
```

### Climatology

Note: colour scale is not consistent between figures.

Figures are not interactive.

::: panel-tabset

```{r}
#| fig-height: 6
#| results: "asis"

for(i in seq_along(counties)){
  
  county.i <- counties[i]
  dat.i <- county_month_depth %>% 
    filter(county == county.i)
  
  cat('\n###', county.i, '\n')
  
  p <- ggplot(
    dat.i,
    aes(month, mean, group = depth, col = factor(depth))
  ) +
    geom_point(size = 3) +
    geom_line(aes(col = factor(depth))) +
    geom_ribbon(
      aes(ymin = mean - stdev, ymax = mean + stdev,
          fill = factor(depth)), alpha = 0.25, col = NA
    ) +
    scale_x_discrete(name = "") +
    scale_y_continuous(
      "Temperature (mean +/- standard deviation)",
      limits = c(-2, 27)
    )

  print(p)
}
  
```

:::
