==> R/temperature_degree_c/3.2_apply_temp_county_rolling_sd_flag.rmd <==
---
format: 
  html:
    toc: true
    toc-depth: 3
    embed-resources: true
editor_options: 
  chunk_output_type: console
params: 
  county: county
  thresh: thresh
---

```{r, set-up, include=FALSE}
# code was modified Nov 16, 2023 to check thresh = q_999
# should be modified again for clarity and speed

knitr::opts_chunk$set(echo = FALSE, dpi = 600, fig.width = 8)

library(data.table)
library(dplyr)
library(DT)
library(ggplot2)
library(here)
library(purrr)
library(qaqcmar)
library(sensorstrings)
library(strings)
library(tgc)
library(tidyr)

source(here("functions/subchunkify.R"))

theme_set(theme_light())
flag_colours <- c("chartreuse4", "#E6E1BC", "#EDA247", "#DB4325", "grey24")

dt_options <- list(
      dom = 'ft',
      paging = FALSE,
      searching = TRUE,
      pageLength = 500,
      columnDefs = list(list(className = 'dt-center', targets = "_all"))
)

thresholds_county <- fread(here("output/temperature_rolling_sd_thresholds.csv")) %>% 
  #pivot_longer(cols = c(-"county"), names_to = "threshold", values_to = "threshold_value") %>% 
  mutate(county = tolower(county)) %>% 
  filter(county == params$county)
 # filter(county == params$county, threshold == params$thresh)

county <- params$county

thresh <- thresholds_county$threshold_value
 
# dat_raw <- readRDS(here("data/temp_rolling_sd_prelim_qc.rds")) %>% 
#   select(-c(sensor_type, int_sample, n_sample, rolling_sd_flag_temperature_degree_c)) %>% 
#   mutate(county = tolower(county)) %>% 
#   filter(county == params$county) 

path <- file.path("R:/data_branches/water_quality/processed_data/deployment_data")

path <- paste0(path, "/", county, "/new")

```

```{r, import-data}
#| message: FALSE

dat <- list.files(path, full.names = TRUE, pattern = ".rds") %>%
  purrr::map(readRDS) %>%
  list_rbind() %>%
  select(
    -c(waterbody, lease, latitude, longitude, string_configuration,
       contains("dissolved_oxygen"), contains("salinity"), contains("uncorrected"),
       contains("sensor_depth_measured")
    )
  )


if(!is.numeric(thresh)) stop(paste0("threshold must be numeric, not: ", thresh))

dat_qc <- dat %>% 
  qc_test_rolling_sd(
    rolling_sd_table = data.frame(
      variable = "temperature_degree_c",
      rolling_sd_max = thresh
    )
  )

# summarise flagged observations
qc_summary <- dat_qc %>%
  rename(rolling_sd_flag = rolling_sd_flag_temperature_degree_c) %>%
  qc_assign_flag_labels() %>%
  group_by(deployment_range, station, rolling_sd_flag) %>%
  summarise(flag_count = n()) %>%
  mutate(flag_percent = round((100 * flag_count / sum(flag_count)), digits = 3)) %>%
  ungroup() %>%
  select(-flag_count) %>%
  pivot_wider(names_from = "rolling_sd_flag", values_from = "flag_percent") %>%
  arrange(station, deployment_range)
```

# `r params$county` QC Flags: Temperature, Rolling standard deviation

`r Sys.Date()`

`r thresh`

# Summary Table

Percent of observations in each deployment assigned with each qc flag. 

```{r}
datatable(qc_summary, options = dt_options, rownames = FALSE)
```

threshold: `r thresh`

# Figures

```{r, body} 
#| results: 'asis'
#| message: FALSE

if("Suspect/Of Interest" %in% colnames(qc_summary)) {
  
  depl_id <- qc_summary %>%
    filter(!is.na(`Suspect/Of Interest`)) %>% 
    distinct(deployment_range, station) 
  
  stations <- unique(depl_id$station)
  
  for(i in seq_along(stations)){
    
    station_i <- stations[i]
    
    deployments <- depl_id %>% filter(station == station_i)
    deployments <- deployments$deployment_range
    
    cat('\n##', station_i, '\n')
    
    cat('\n')
    
    # plot for each deployment
    for(j in 1:length(deployments)){
      
      depl_j <- deployments[j]
      
      # subset data to station of interest
      dat_j <- dat_qc %>%
        filter(station == station_i, deployment_range == depl_j) %>%
        qc_pivot_longer(qc_tests = "rolling_sd") %>% 
        ss_convert_depth_to_ordered_factor()
      
      # height of figure j
      n_sensor <- length(unique(dat_j$sensor_serial_number))
      
      if(n_sensor == 1) h = 2.75
      if(n_sensor > 1) h = 2.75 * n_sensor
      
      
      p <- qc_plot_flags(dat_j, qc_tests = "rolling_sd")
      p <- p$temperature_degree_c$rolling_sd

      cat('\n###', depl_j, '\n')
      
      subchunkify(p, fig_height = h, fig_width = 8.5)
      
      #cat(paste("Figure ", k, ": ", station.j, fig.caption,  sep = ""), '\n')
      
      cat('\n')
    }
  }
}

```
