==> R/temperature_degree_c/4.1_temp_spike.qmd <==
---
format: 
  html:
    toc: true
    toc-depth: 3
    embed-resources: true
editor_options: 
  chunk_output_type: console
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, dpi = 600, fig.width = 8)

library(dplyr)
library(DT)
library(here)
library(ggplot2)
library(leaflet)
library(lubridate)
library(plotly)
library(purrr)
library(qaqcmar)
library(RColorBrewer)
library(readr)
library(sensorstrings)
library(strings)
library(summaryplots)
library(tidyr)
library(viridis)

source(here("functions/subchunkify.R"))

theme_set(theme_light())

x_axis <- scale_x_continuous("temperature (deg C)")

dt_options <- list(
  dom = 'ft',
  paging = FALSE,
  searching = TRUE,
  scrollY = "500px",
  columnDefs = list(list(className = 'dt-center', targets = "_all"))
)

# data - preliminary QC 
dat_raw <- readRDS(here("data/temp_rolling_sd_prelim_qc.rds")) %>% 
  select(
    -c(sensor_type, int_sample, n_sample, rolling_sd_flag_temperature_degree_c)
  ) %>% 
  mutate(
    depth = factor(sensor_depth_at_low_tide_m),
    year = factor(year(timestamp_utc)),
    month = month(timestamp_utc)
  ) %>% 
  na.omit() 

counties <- unique(dat_raw$county)

dat_all <- NULL

for(i in seq_along(counties)) {
  
  county_i <- counties[i]
  
  dat_all[[i]] <- dat_raw %>% 
    filter(county == county_i) %>% 
    qc_test_spike(keep_spike_cols = TRUE) %>% 
    select(-c(spike_high, spike_low, spike_flag_temperature_degree_c))
}

dat_all <- dat_all %>% 
  map_df(bind_rows) %>% 
  # filter_out_suspect_obs() and other foos expects data in the OLD format
  select(
    COUNTY = county, 
    STATION = station, 
    DEPLOYMENT_PERIOD = deployment_range,
    TIMESTAMP = timestamp_utc,
    VALUE = value_temperature_degree_c,
    DEPTH = sensor_depth_at_low_tide_m,
    everything()) %>% 
  mutate(
    VARIABLE = "Temperature", 
    UNITS = "degree C", 
    YEAR = factor(year(TIMESTAMP))
  ) %>% 
  filter(!is.na(spike_value)) %>% 
  convert_depth_to_ordered_factor()

# filtered data
dat_filt <- dat_all %>% 
  filter(
    !(STATION %in% c("Piper Lake", "Hourglass Lake", "0193", "Sissiboo")),
    !(COUNTY == "Inverness" & DEPTH %in% c(18, 23, 26, 28, 36, 40)) 
  )
     

county_pal <- get_county_colour_palette(length(unique(dat_all$COUNTY)))
depth_pal <- viridis(length(unique(dat_all$DEPTH)), direction = -1)
year_pal <- viridis( length(unique(dat_all$YEAR)), option = "C")

theme_facet_plotly <- theme(
  panel.spacing.y = unit(30, "lines"),
  panel.spacing.x = unit(20, "lines")
)

```

# Temperature Spikes

November 14, 2023

- Data submitted to the Open Data Portal in December 2022.
- Depth rounded to nearest whole number
  - Results in depth of 0 m for 1042, Captains Pond, Ile due Havre, Sober Island, and Back Harbour.
  
- Preliminary QC (submitted to Open Data Portal)
  - Obvious outliers removed.

- Filtered Data
  - Fresh water and other "outlier" stations not considered.
    - Piper Lake, Hourglass Lake, 0193 (Denas Pond), Sissiboo
  - Data below the thermocline in Whycocomagh Bay
    - 0814x East and 0814x West, Whycocomagh Basin: 18 m
    - Aberdeen, Whycocomagh Bay: 18, 23, and 28 m
    - Deep Basin, Whycocomagh Basin: 26, 36, and 40 m

- Following QARTOD: $$spike_{value} = abs(value - spike_{ref})$$ where $spike_{ref} = (lead_{value} - lag_{value}) / 2$

## Preliminary QC 

### Pooled

::: panel-tabset

#### Figure 1: All observations
```{r}
#| label: fig-dist-all
#| fig-height: 6
#| fig-cap: Distribution of temperature observations (binwidth = 0.1 degree C).

p <- plot_histogram(dat_all, hist_col = spike_value, binwidth = 0.1) +
  x_axis

ggplotly(p)
```

#### Figure 2: By depth
```{r}
#| label: fig-dist-all-depth
#| fig-height: 6
#| fig-cap: Distribution of temperature observations (binwidth = 0.1 degree C), coloured by depth.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1,
  fill_col = DEPTH, pal = depth_pal
) +
  x_axis

ggplotly(p)
```

#### Figure 3: By year
```{r}
#| label: fig-dist-all-year
#| fig-height: 6
#| fig-cap: Distribution of temperature observations (binwidth = 0.1 degree C), coloured by year.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1, 
  fill_col = YEAR, pal = year_pal
) +
  x_axis

ggplotly(p)
```

:::

### County

::: panel-tabset

#### Figure 4: County 
```{r}
#| label: fig-dist-county
#| fig-height: 10
#| fig-cap: Distribution of temperature observations by county (binwidth = 0.1 degree C).

p <- dat_all %>% 
  plot_histogram(hist_col = spike_value, binwidth = 0.1) +
  x_axis +
  facet_wrap(~COUNTY, scales = "free", ncol = 3) +
  theme_facet_plotly

ggplotly(p)
```

#### Figure 5: County + Depth
```{r}
#| label: fig-dist-county-depth
#| fig-height: 10
#| fig-cap: Distribution of temperature observations (binwidth = 0.1 degree C), within each county, coloured by depth.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1, fill_col = DEPTH, pal = depth_pal
) +
  x_axis +
  facet_wrap(~COUNTY, scales = "free", ncol = 3) +
  theme_facet_plotly
  
ggplotly(p)
```

#### Figure 6: County + Year
```{r}
#| label: fig-dist-county-year
#| fig-height: 10
#| fig-cap: Distribution of temperature observations (binwidth = 0.1 degree C), within each county coloured by year.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1, fill_col = YEAR, pal = year_pal
) +
  x_axis +
  facet_wrap(~COUNTY, scales = "free", ncol = 3) +
  theme_facet_plotly

ggplotly(p)
```
:::

#### County + Month + Year

::: panel-tabset
```{r}
#| fig-height: 10
#| results: "asis"

counties <- unique(dat_all$COUNTY)

dat_all <- dat_all %>% 
  mutate(MONTH = month(TIMESTAMP, label = TRUE))

for(i in seq_along(counties)){
  
  county_i <- counties[i]
  dat_i <- dat_all %>% 
    filter(COUNTY == county_i) 
  
  cat('\n###', county_i, '\n')
  
  p <- plot_histogram(
    dat_i, hist_col = spike_value, binwidth = 0.1, fill_col = YEAR, pal = year_pal
  ) +
    x_axis +
    facet_wrap(~MONTH, ncol = 3)  
  
  print(p)
}

```
:::

### Summary Statistics

::: panel-tabset

#### Figure 7: Central Tendancy
```{r}
#| fig-height: 6
#| fig-cap: The mean and standard deviation of spike value in each county.
#| label: fig-mean-all

stats_all <- dat_all %>% 
  summarise(
    max = max(spike_value),
    mean = mean(spike_value),
    med = median(spike_value),
    stdev = sd(spike_value),
    
    q_90 = quantile(spike_value, prob = 0.90),
    q_95 = quantile(spike_value, prob = 0.95),
    q_997 = quantile(spike_value, prob = 0.997)
  ) %>% 
  mutate(COUNTY = "All") %>% 
  bind_rows(
    dat_all %>% 
      group_by(COUNTY) %>% 
      summarise(
        max = max(spike_value),
        mean = mean(spike_value),
        med = median(spike_value),
        stdev = sd(spike_value),
        
        q_90 = quantile(spike_value, prob = 0.90),
        q_95 = quantile(spike_value, prob = 0.95),
        q_997 = quantile(spike_value, prob = 0.997)
      ) 
  ) %>%
  mutate(
    variable = "temperature_degree_c",
    thresh_3sd = mean + 3 * stdev,
    across(.cols = where(is.numeric), .fns = round, digits = 2)
  ) %>% 
  select(county = COUNTY, everything())

p <- plot_mean_sd_county(stats_all)

ggplotly(p)
```

#### Figure 8: Quantiles
```{r}
#| fig-cap: Upper quantiles.
#| fig-height: 6
#| label: fig-quantiles-all

p <- stats_all %>% 
  select(county, contains("q")) %>% 
  pivot_longer(cols = contains("q"), names_to = "quantile") %>% 
  ggplot(aes(quantile, value, group = county, col = county)) +
  geom_point() +
  geom_line() +
  scale_color_manual(values = c("grey20", county_pal), drop = FALSE) +
  scale_y_continuous(limits = c(0, 2))

ggplotly(p)
```

#### Table 1: Summary statistics
```{r}
stats_all %>% 
  select(-c(variable, max, contains("thresh"))) %>% 
  datatable(
    options = list(
      dom = 'ft',
      paging = FALSE,
      searching = TRUE,
      columnDefs = list(list(className = 'dt-center', targets = "_all"))
    ), rownames = FALSE)
```
:::




## Filtered Data

### Pooled

::: panel-tabset

#### Figure 9: All observations
```{r}
#| fig-height: 6
#| fig-cap: Distribution of temperature observations (binwidth = 0.1 degree C).

p <- plot_histogram(dat_filt, hist_col = spike_value, binwidth = 0.1) +
  x_axis

ggplotly(p)
```

#### Figure 10: By depth
```{r}
#| fig-height: 6
#| fig-cap: Distribution of temperature observations (binwidth = 0.1 degree C), coloured by depth.

p <- plot_histogram(
  dat_filt, hist_col = spike_value, binwidth = 0.1,
  fill_col = DEPTH, pal = depth_pal
) +
  x_axis

ggplotly(p)
```

#### Figure 11: By year
```{r}
#| fig-height: 6
#| fig-cap: Distribution of temperature observations (binwidth = 0.1 degree C), coloured by year.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1, 
  fill_col = YEAR, pal = year_pal
) +
  x_axis

ggplotly(p)
```

:::

### County

::: panel-tabset

#### Figure 12: County 
```{r}
#| fig-height: 10
#| fig-cap: Distribution of temperature observations by county (binwidth = 0.1 degree C).

p <- dat_filt %>% 
  plot_histogram(hist_col = spike_value, binwidth = 0.1) +
  x_axis +
  facet_wrap(~COUNTY, scales = "free", ncol = 3) +
  theme_facet_plotly

ggplotly(p)
```

#### Figure 13: County + Depth
```{r}
#| fig-height: 10
#| fig-cap: Distribution of temperature observations (binwidth = 0.1 degree C), within each county, coloured by depth.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1, fill_col = DEPTH, pal = depth_pal
) +
  x_axis +
  facet_wrap(~COUNTY, scales = "free", ncol = 3) +
  theme_facet_plotly
  
ggplotly(p)
```

#### Figure 14: County + Year
```{r}
#| fig-height: 10
#| fig-cap: Distribution of temperature observations (binwidth = 0.1 degree C), within each county coloured by year.

p <- plot_histogram(
  dat_all, hist_col = spike_value, binwidth = 0.1, fill_col = YEAR, pal = year_pal
) +
  x_axis +
  facet_wrap(~COUNTY, scales = "free", ncol = 3) +
  theme_facet_plotly

ggplotly(p)
```
:::

#### County + Month + Year

::: panel-tabset
```{r}
#| fig-height: 10
#| results: "asis"

dat_filt <- dat_filt %>% 
  mutate(MONTH = month(TIMESTAMP, label = TRUE))

for(i in seq_along(counties)){
  
  county_i <- counties[i]
  dat_i <- dat_filt%>% 
    filter(COUNTY == county_i) 
  
  cat('\n###', county_i, '\n')
  
  p <- plot_histogram(
    dat_i, hist_col = spike_value, binwidth = 0.1, fill_col = YEAR, pal = year_pal
  ) +
    x_axis +
    facet_wrap(~MONTH, ncol = 3)  
  
  print(p)
}

```
:::

### Summary Statistics

::: panel-tabset

#### Figure 15: Central Tendancy
```{r}
#| fig-height: 6
#| fig-cap: The mean and standard deviation of spike value in each county.
#| label: fig-mean-filt

stats_filt <- dat_filt %>% 
  summarise(
    max = max(spike_value),
    mean = mean(spike_value),
    med = median(spike_value),
    stdev = sd(spike_value),
    
    q_90 = quantile(spike_value, prob = 0.90),
    q_95 = quantile(spike_value, prob = 0.95),
    q_997 = quantile(spike_value, prob = 0.997)
  ) %>% 
  mutate(COUNTY = "All") %>% 
  bind_rows(
    dat_filt %>% 
      group_by(COUNTY) %>% 
      summarise(
        max = max(spike_value),
        mean = mean(spike_value),
        med = median(spike_value),
        stdev = sd(spike_value),
        
        q_90 = quantile(spike_value, prob = 0.90),
        q_95 = quantile(spike_value, prob = 0.95),
        q_997 = quantile(spike_value, prob = 0.997)
      ) 
  ) %>%
  mutate(
    variable = "temperature_degree_c",
    thresh_3sd = mean + 3 * stdev,
    across(.cols = where(is.numeric), .fns = round, digits = 2)
  ) %>% 
  select(county = COUNTY, everything())

p <- plot_mean_sd_county(stats_filt)

ggplotly(p)
```

#### Figure 16: Quantiles
```{r}
#| fig-cap: Upper quantiles.
#| fig-height: 6
#| label: fig-quantiles-filt

p <- stats_filt %>% 
  select(county, contains("q")) %>% 
  pivot_longer(cols = contains("q"), names_to = "quantile") %>% 
  ggplot(aes(quantile, value, group = county, col = county)) +
  geom_point() +
  geom_line() +
  scale_color_manual(values = c("grey20", county_pal), drop = FALSE) +
  scale_y_continuous(limits = c(0, 2))

ggplotly(p)
```

#### Table 2: Summary statistics
```{r}
stats_filt %>% 
  select(-c(variable, max, contains("thresh"))) %>% 
  datatable(
    options = list(
      dom = 'ft',
      paging = FALSE,
      searching = TRUE,
      columnDefs = list(list(className = 'dt-center', targets = "_all"))
    ), rownames = FALSE)
```
:::
